mcp_finetune_dataset/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.39.0
scipy>=1.10.0
numpy>=1.24.0
requests>=2.28.0
pyyaml>=6.0
tqdm>=4.64.0
huggingface_hub>=0.16.0
wandb>=0.15.0
safetensors>=0.3.0
sentencepiece>=0.1.99
protobuf>=3.20.0
tokenizers>=0.13.0
psutil>=5.9.0
gitpython>=3.1.0